feat: move checking storeId logic(mainActivity -> homeFragment)
