+++ feature/main/src/main/java/org/swm/att/home/MainViewModel.kt
@@ -1,38 +1,22 @@
 package org.swm.att.home
 
-import android.content.Context
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import org.json.JSONObject
 import org.swm.att.common_ui.presenter.base.BaseViewModel
-import org.swm.att.common_ui.state.UiState
-import org.swm.att.common_ui.util.Formatter
 import org.swm.att.common_ui.util.JWTUtils
 import org.swm.att.common_ui.util.JWTUtils.unixTimeToDateTime
-import org.swm.att.common_ui.util.getUTCDateTime
-import org.swm.att.domain.entity.HttpResponseException
-import org.swm.att.domain.entity.request.StoreVO
-import org.swm.att.domain.entity.response.PreorderVO
-import org.swm.att.domain.entity.response.StoreIdVO
-import org.swm.att.domain.repository.AttOrderRepository
 import org.swm.att.domain.repository.AttPosUserRepository
 import org.swm.att.home.constant.NavDestinationType
-import org.swm.att.home.util.alarm.AlarmManager
-import java.util.Calendar
 import javax.inject.Inject
 
 @HiltViewModel
 class MainViewModel @Inject constructor(
     private val attPosUserRepository: AttPosUserRepository,
-    private val attOrderRepository: AttOrderRepository,
-    @ApplicationContext private val context: Context
 ) : BaseViewModel() {
     private val _refreshExist = MutableLiveData<Boolean>()
     val refreshExist: LiveData<Boolean> = _refreshExist
@@ -42,14 +26,6 @@ class MainViewModel @Inject constructor(
     val curSelectedScreen: LiveData<NavDestinationType> = _curSelectedScreen
     private val _isGlobalAction = MutableLiveData(false)
     val isGlobalAction: LiveData<Boolean> = _isGlobalAction
-
-    private val todayPreorderList = arrayListOf<PreorderVO>()
-    private var page = 1
-
-    private val _storeIdExist = MutableLiveData<Int>()
-    val storeIdExist: LiveData<Int> = _storeIdExist
-    private val _registerStoreState = MutableStateFlow<UiState<StoreIdVO>>(UiState.Loading)
-    val registerStoreState: StateFlow<UiState<StoreIdVO>> = _registerStoreState
 
     fun checkRefreshToken() {
         viewModelScope.launch {
@@ -64,39 +40,6 @@ class MainViewModel @Inject constructor(
                     refreshToken(curRefreshToken)
                 }
             }
-        }
-    }
-
-    fun checkStoreId() {
-        viewModelScope.launch(attExceptionHandler) {
-            val storeId = attPosUserRepository.getStoreId()
-            Log.d("MainViewModel", "checkStoreId: $storeId")
-            if (storeId == -1) {
-                // TODO 새로운 가게 등록 화면으로 전환 -> _storeIdExist 활용
-                // 임시 token을 활용해 가게 바로 등록
-                attPosUserRepository.registerStore(
-                    StoreVO(
-                        name = "temp",
-                        address = "temp",
-                        openingHours = emptyList()
-                    )
-                ).collect { result ->
-                    result.onSuccess {
-                        _registerStoreState.value = UiState.Success(it)
-                    }.onFailure {
-                        val errorMsg = if (it is HttpResponseException) it.message else "가게 등록 실패"
-                        _registerStoreState.value = UiState.Error(errorMsg)
-                    }
-                }
-            }
-
-        }
-    }
-
-    fun setStoreId(storeId: Int) {
-        viewModelScope.launch(attExceptionHandler)  {
-            Log.d("MainViewModel", "setStoreId: $storeId")
-            attPosUserRepository.saveStoreId(storeId)
         }
     }
 
@@ -143,44 +86,5 @@ class MainViewModel @Inject constructor(
 
     fun isDestinationDiff(destination: NavDestinationType): Boolean {
         return curSelectedScreen.value != destination
-    }
-
-    fun getTodayPreorder() {
-        viewModelScope.launch(attExceptionHandler) {
-            val date =
-                Formatter.getStringByDateTimeBaseFormatter(Calendar.getInstance().time.getUTCDateTime())
-            val storeId = attPosUserRepository.getStoreId()
-            attOrderRepository.getPreOrders(storeId, page, date).collect { result ->
-                result.onSuccess {
-                    todayPreorderList.addAll(it.preOrders)
-                    page += 1
-                    if (page < it.lastPage) {
-                        getTodayPreorder()
-                    } else {
-                        setPreorderAlarm()
-                    }
-                }.onFailure {
-                    val errorMsg = if (it is HttpResponseException) it.message else "예약 내역 불러오기 실패"
-                    Log.e("getTodayPreorder", errorMsg!!)
-                }
-            }
-        }
-    }
-
-    private fun setPreorderAlarm() {
-        for (preorderItem in todayPreorderList) {
-            // 각 아이템에 대한 알람 시간과 기타 설정 가져옴
-            AlarmManager.setPreorderAlarm(
-                context,
-                preorderItem.orderedFor,
-                preorderItem.phone,
-                preorderItem.totalCount,
-                preorderItem.id
-            )
-        }
-    }
-
-    fun cancelAllPreorderAlarm() {
-        AlarmManager.cancelAllAlarm(context)
     }
 }
